Fix runtime error: change log interval types from unsigned int to double
Runtime Error Fixed:
- Line 162: "invalid argument type: found 'pointer to unsigned int', expected 'pointer to double'"

Type Changes in OverchargeParams struct:
- logIntervalSlow: unsigned int → double
- logIntervalFast: unsigned int → double

Reason:
GetCtrlVal() expects a pointer to double for numeric controls in LabWindows/CVI.
The UI controls are defined as numeric/double types, not integer types.

Using double is also more consistent with other timing parameters in the struct
(chargeDurationMinutes, eisIntervalSlow_minutes, etc.) and allows fractional seconds.

+++ exp_overcharge.h
@@ -107,8 +107,8 @@ typedef struct {
     int pauseChargeDuringEIS;       // 1 = pause, 0 = continue
 
     // Logging (adaptive)
-    unsigned int logIntervalSlow;   // Data logging interval when SOC < threshold (seconds)
-    unsigned int logIntervalFast;   // Data logging interval when SOC >= threshold (seconds)
+    double logIntervalSlow;   // Data logging interval when SOC < threshold (seconds)
+    double logIntervalFast;   // Data logging interval when SOC >= threshold (seconds)
 } OverchargeParams;
 
 // Temperature data point (reuse from exp_temp_ramp)
